errorhandle commit

+++ main.py
@@ -15,7 +15,7 @@ mongo = PyMongo(app)
 def add():
     name = request.json['name']
     pwd = request.json['pass']
-    if name and pwd and request.method=="POST":
+    if name and pwd:
         id = mongo.db.emp.insert_one({'name':name,'pass':pwd})
         resp = "User added Succesfully"
         return jsonify(resp)
@@ -23,41 +23,54 @@ def add():
     else:
         return not_found()
     
-@app.errorhandler(404)
-def not_found(error=None):
-    message = {
-        'status':404,
-        'message':"Not Found" + request.url
-    }
-    resp = jsonify(message)
-    return resp
+
 
 @app.route('/all-emp',methods=['GET'])
 def show():
-    users = mongo.db.emp.find()
-    resp = dumps(users)
-    return resp
+        users = mongo.db.emp.find()
+        resp = dumps(users)
+        return resp
 
 
 @app.route('/show-emp/<id>',methods=['GET'])
 def show_one(id):
-    users = mongo.db.emp.find_one({'_id':ObjectId(id)})
-    resp = dumps(users)
-    return resp
+    if len(str(id))== 24:
+        users = mongo.db.emp.find_one({'_id':ObjectId(id)})
+        if users != None:
+            resp = dumps(users)
+            return resp
+    return "No User Found with id : %s"%id
 
 @app.route('/dlt-emp/<id>',methods=['DELETE'])
 def dlt_one(id):
-    mongo.db.emp.delete_one({'_id':ObjectId(id)})
-    resp = "User deleted Succesfully"
-    return jsonify(resp)
+    if len(str(id))== 24:
+        users = mongo.db.emp.find_one({'_id':ObjectId(id)})
+        if users != None:
+            mongo.db.emp.delete_one({'_id':ObjectId(id)})
+            resp = "User deleted Succesfully"
+            return jsonify(resp)
+    return "No User Found with id : %s"%id
 
 @app.route('/update-emp/<id>',methods=['PUT'])
 def update(id):
-    name = request.json['name']
-    pwd = request.json['pass']
-    mongo.db.emp.update_one({'_id':ObjectId(id)},{'$set':{'name':name,'pass':pwd}})
-    resp = "User updated Succesfully"
-    return jsonify(resp)
+    if len(str(id))== 24:
+        users = mongo.db.emp.find_one({'_id':ObjectId(id)})
+        if users != None:
+            name = request.json['name']
+            pwd = request.json['pass']
+            mongo.db.emp.update_one({'_id':ObjectId(id)},{'$set':{'name':name,'pass':pwd}})
+            resp = "User updated Succesfully"
+            return jsonify(resp)
+    return "No User Found with id : %s"%id
+
+@app.errorhandler(404)
+def not_found(error=None):
+    message = {
+        'status':404,
+        'message':"Not Found " + request.url
+    }
+    resp = jsonify(message)
+    return resp
 
 if __name__ == "__main__":
     app.run()
